One potential problem that I fixed was an automatic array whose size was being calculated on the declaration line. I moved the allocation of the array to the C wrapper, and will shortly fix it in the Fortran routine.
Unfortunately, I'm still getting a core dump when the input values to
cancor are floating point. It works fine if the input is double
precision. The problem is not very reproducible, as the
code works fine on bluesky, MacOSX, and in the debug version on
tempest (not the optimized version). It core dumps with both the
debug and optimized versions on a Sun.


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@8480 10fe2890-8a40-4696-8b1d-7ccbb19c398e