This adds support for the vector rotation variables for Gds grid types 3 - lambert conformal 5 - polar stereographic 201, 202, and 203: the arakawa rotated lat/lon grids.
Note that regular lat/lon and gaussian grids do not have rotated vectors, so
the rotation variable does not need to be implemented for these.

I do not have any examples of GDS types 201 or 202, but I implemented
support for them anyway, since there are ncep codes to do these.

I am still working on an implementation for a regular rotated lat/lon grid (GDS type 10).
I need an example.


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@8490 10fe2890-8a40-4696-8b1d-7ccbb19c398e