Based upon a user request, I added another attribute to be returned by regCOef/regline. [ "IDL does this why doesn't NCL."]
This is "rstd" ... the standard deviation of the
regression coefficient.

       r@rstd

The fortran changes were minimal since this was
already computed ... just not returned.

Yes, I realize that this is for a034.

MH just needs to duplicate the 'tval' stuff in the interface.

I will change the .xml files when MH says it is ok.

---
I did show the user how to calculate the rstd from the information
aleady returned. This is not obvious however.


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@8579 10fe2890-8a40-4696-8b1d-7ccbb19c398e