Added support for the NC_FORMAT_NETCDF4_CLASSIC netcdf file format when NCL is compiled with the NetCDF4 library. This format enforces the "classic" NetCDF interface but stores the data in an HDF5-format file. This format does support zlib file compression, which is accessible via a new file option called "CompressionLevel". "CompressionLevel" is an integer-valued option with a valid range from -1 to 9, where -1 means turn off compression, and 9 is the maximum level of compression. Higher levels of compression produce smaller files but take more time to write. The default is -1. Note that this option and all NetCDF4 library-dependent code is conditionally-compiled based on dectection of a NetCDF4-only #-defined constant. Eventually when the NetCDF4 library is released in a stable form the conditional compilation stuff will be removed. Currently the NetCDF4 library has a bug that results in variable attributes not always being written to the file properly, so it is not yet ready for prime-time.
I also added a new logical NetCDF option called "MissingToFillValue". This
option only has an effect if set before opening the file. When
set True, it causes a "virtual" _FillValue attribute to be created for
any variable in the file that has a "missing_value" attribute but not
a "_FillValue" attribute. The purpose is to more gracefully handle files
that use the COARDS-compliant "missing_value" instead of "_FillValue" to indicate missing data.
Note that if a variable in a file has both a "missing_value" and a "_FillValue",
or if it has neither the option does nothing.
The virtual _FillValue attribute is not actually part of
the NetCDF file, but appears to be from within the NCL environment.
It is propagated upon assignment of the variable just like any other
attribute. If the file is opened for writing and you assign to the attribute, it
becomes an actual attribute. Therefore to make it a permanent part of a file
you could say, for example:

f->v@_FillValue = f->v@_FillValue

Unfortunately, there is no simple way within NCL to tell whether a particular
instance of a _FillValue attribute (where there is also a missing_value attribute
with the same value) is virtual or actual. The only way to tell would be to
open the file with opposite settings of the option and see whether the
attribute was present in one case and not the other. However note that the
'f->v@_FillValue = f->v@_FillValue' construct would be harmless if the attribute
actually exists in the file already.

The default value of the option is True.


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@8684 10fe2890-8a40-4696-8b1d-7ccbb19c398e