A variable set in a DATA statement may be changed.
The user may specify that the changed value be SAVE'd

If the SAVE statement is not used, the fortran standard allows
the behavior to be unspecified. IE: the compiler writer is allowed
to do 'whatever'.

--
I did not use the SAVE statement but did comment out the

c djs DATA CHECK /0.D0/

c djs IF (CHECK.eq.0.d0) then
c djs END IF

I ran a .so check and it works fine.


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@8709 10fe2890-8a40-4696-8b1d-7ccbb19c398e