Well, the Anemone won this battle. He wanted "dtrend" to return a _FillValue attribute even though the output values are guaranteed not to contain any missing values.
Here was his reasoning on it, for future reference:

I always try to set _FillValue, regardless of whether or not my data has
any missing data in it. I thought that this was a good NCL programming
practice. I know that in some of my programs, I will dtrend the data,
and then mask the data, setting all land values for instance to
_FillValue. If _FillValue is now not returned from dtrend then my
current script will fail.

I don't want his scripts or anyone else's scripts to fail, for that matter,
so be it.

--M


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@8716 10fe2890-8a40-4696-8b1d-7ccbb19c398e