Made a minor change inside the wrapper for c_tditri in the calling sequence to this routine.
In ripW.c, added code that will search for the psadilookup file in
a standard location, and then in a location specified by the PSADILOOKUP_PATH
environment variable.

Also added some more checking of the input variable sizes. Here's what they
can be, according to an email from Julie Schramm:

There are 3 physically realistic combinations of dimensions:

1) p,t,q,z (time, nlev, nlat nlon), and psfc,zsfc(time,nlat,nlon)

2) p,t,q,z (nlev, nlat nlon), and psfc,zsfc(nlat,nlon)
This is the standard case for doing the calculation in each column over
a grid.

3) p,t,q,z (nlev), and psfc,zsfc scalars
This would be the case for doing the calculation in a single column.


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@8789 10fe2890-8a40-4696-8b1d-7ccbb19c398e