Fix for problem on some systems where, if environment variables NCL_DEF_LIB_DIR or NCL_DEF_SCRIPTS_DIR are not set, NCL would crash (segmentation fault or bus error).  In each case:  if the environment variable was set, but set to a directory that doesn't exist or is not accessible by the user, it cannot be successfully opened (via opendir()).  In that case, it's not necessary to attempt to close that directory (via closedir()) so that call is removed.
On some systems, closedir() is sufficiently robust to ignore its argument if that
argument is NULL.  But this isn't true on all systems.  Since it's not necessary
to close a directory that was never opened in the first place, removing this call
should be safe.

Also changed the error message emitted when either NCL_DEF_LIB_DIR or
NCL_DEF_SCRIPTS_DIR is set, but set to a bogus value.  NCL was emitting a "fatal"
error message, but continuing on anyway.  Changed error message to a "warning"
level.


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@8841 10fe2890-8a40-4696-8b1d-7ccbb19c398e