Added two new resources to gsn_panel that allow you to individually change the X and/or Y positions of individual plots in a panel: gsnPanelXF and gsnPanelYF.
This should hopefully appease some of the folks that want to create
a panel of slightly different sized plots, and then need to scoot one
or more up or down.

The new resources must be arrays of
the same length as the number of plots being paneled. Any plots whose
X,Y positions you don't want to change, just use "-1".  For example,
if you want to create a panel of four plots with 2 rows and 2 columns,
and you need to move the second column of plots closer to the first column,
you might use something like this:

   pnlres@gsnPanelXF = (/-1,0.5,-1,0.5/)
   gsn_panel(wks,plots,(/2,2/),pnlres)

This will cause the internally-calculated X positions to be used for
plots 0 and 2 (the leftmost plots), and to use 0.5 as the X position
for plots 1 and 3 (the rightmost plots).


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@8886 10fe2890-8a40-4696-8b1d-7ccbb19c398e