strlen() updated to account for missing values.  Any string that is "missing" (not assigned) is given a length equal to the default integer missing value (-999), and the _FillValue attribute is set.  If, for some reason, a user deletes the _FillValue attribute associated with a string that is "missing" is assigned a length of "7" (the length of the string "missing").
Strings originally created with no assigned value (as with "new()") but later
assigned a value will result in the _FillValue attribute not being assigned to
the resultant length associated with the string.  Any array of strings that
contain even one missing value will result in the array of length of strings
having a _FillValue attribute (equal to the default integer missing value, -999).


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@8961 10fe2890-8a40-4696-8b1d-7ccbb19c398e