Repaired references to an unitialized variable in _NclIabs.
NclGRIB: Fixed a error that resulted from creating the ensembleX_info
variable as a 1DCoordinate variable type rather than just as a regular
variable. It's not a coordinate variable and since it's not monotonic,
you can't pretend it is.

NclHDF.c: was barfing when you tried to create a 'long' type variable.
Fixed.


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@8971 10fe2890-8a40-4696-8b1d-7ccbb19c398e