Version 3.1 of Spherepack did not fix the problem in shsgc as we originally thought. The code:
      abel(imn)=sqrt(float((2*n+1)*(m+n-2)*(m+n-3))/
     1               float(((2*n-3)*(m+n-1)*(m+n))))
      bbel(imn)=sqrt(float((2*n+1)*(n-m-1)*(n-m))/
     1               float(((2*n-3)*(m+n-1)*(m+n))))
      cbel(imn)=sqrt(float((n-m+1)*(n-m+2))/
     1               float(((n+m-1)*(n+m))))

is still returning NaNs if M or N are too large.

Paul Swartztrauber recommended setting FM=float(M) and FN=float(N) and
replacing M and N in the code above with FM and FN. This fixes the
problem, and there are probably more of these that we need to look for
and fix.


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@8974 10fe2890-8a40-4696-8b1d-7ccbb19c398e