Added support for specifying outline boundaries to mask. Added two new resources: mpMaskOutlineSpecifiers and mpOutlineMaskingOn
As with the corresponding fill masking resources, mpOutlineMaskingOn is
by default set True when mpMaskOutlineSpecifiers is set.

These resources allow the user easily to get rid of "inland water" with
both the low res and medium res map databases. With the low res set it
is also very easy to get rid of islands. With the medium res set it is
harder because there are no overall "island" group.

I will send out an example script illustrating these points.


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@8982 10fe2890-8a40-4696-8b1d-7ccbb19c398e