This implements the capability of creating variables without any _FillValue using the NCL 'new' operator.
If the special string "No_FillValue" (case is significant) is passed as
the 3rd argument to 'new', then raw unfilled memory will be returned for
the memory space of the variable to which 'new' is assigned. If the
variable was previously undefined, then it will not be assigned a _FillValue
attribute. However, if it was previously defined and had a _FillValue
attribute, the attribute and its value will be unchanged even though the
elements of the array now have all undefined values.


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@8993 10fe2890-8a40-4696-8b1d-7ccbb19c398e