Modifications that eliminate any fixed length for NCL strings. Strings can now be as long as needed (within the memory limitations of the machine).
Removed the warning messages from the 'strlen' function, and
reworked 'asciiread' to allow unlimited string lengths in a file.
Also modified the string type converter to accept unlimited sizes for
the last dimension of character variables.

Note that symbols (variable, dimension, and attribute names) still are
limited in length to 256.


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@8995 10fe2890-8a40-4696-8b1d-7ccbb19c398e