The wrfW.c was calling a routine that doesn't exist (it was misspelled).
wrapper.py is a python program that will help me wrap a Fortran routine
that is to be built into NCL. It needs improvement, but it works
well for most cases. It knows how to handle routines that have input
arguments of variable dimensions, and it handles cases where numeric
input needs to be coerced to double.

It doesn't handle missing values yet.


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@9049 10fe2890-8a40-4696-8b1d-7ccbb19c398e