Fixed problem with single element coordinate subscripting that Adam found. This was a problem for monotonically decreasing coordinate variables. The test was turned around such that the furthest away coordinate was chosen rather than the closest.
I also fixed an inconsistency for the case of specifying a subscript exactly
half way between 2 coordinates. Now the subscript is chosen in a manner similar
to the round function: the coordinate furthest from zero is chosen: e.g. if coordinates
are 8. and 7. subscript 7.5 chooses 8.; if coordinates are -8. and -7. subscript -7.5
chooses -8.. This is regardless of whether the coordinates are increasing or
decreasing. Also if coordinates are -1. and 1. then subscript 0. chooses
coordinate 1.


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@9060 10fe2890-8a40-4696-8b1d-7ccbb19c398e