In the process of wrapping Dennis' wgt_vert_avg_beta routine, I noticed some problems with the wrapper.py script which I fixed.
Also, I'm checking in the initial version of this wrapper, even though
I've been unable to test it because I can't get at the test data file.
Maybe tomorrow.

Dennis, FYI: the wgt_vert_avg_beta documentation appears to be on the
web, but the corresponding xml file hasn't been checked in with
CVS yet.


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@9089 10fe2890-8a40-4696-8b1d-7ccbb19c398e