This checkin adds the procedure 'loadscript( nclscript : string)', which differs from the load command in the following ways: 1) it accepts string variable referencing the filename as opposed to simply a string literal 2) it can be executed conditionally
One possibly non-intuitive thing to note is that if the script loaded via
loadscript has any proc/funcs, they cannot be accessed within a block that
contains the loadscript procedure. That if testscript.ncl contains procedure
'proc':
if (doit .eq. True) then
  loadscript("testscript.ncl")
  proc()        ; gives an error
end if
as does
begin
  loadscript("testscript.ncl")
  proc()        ; gives an error
end
That is, of course, because the script is not parsed until the whole block
is checked for syntax, and therefore 'proc' is undefined within the block.
'proc' does become available after the 'end if' (assuming 'doit' is True)
or 'end' statements in the examples above.

I also modified the parser (ncl.y) to eliminate an error recovery failure
which has been present in NCL for a long time: if a statement consisted of
an undefined procedure with no arguments NCL was printing the message:
"syntax error: possibly an undefined procedure"  and then failing to process
any more commands. In interactive mode you would get new prompts and you could
enter commands, but no further commands were executed. In batch mode, the
effect was as if NCL had just quit on the line, but in fact in continued to
read commands, but because it had not recovered from the error, it just
threw them away.
In other words both:

undefproc()

and

undefproc(argument)

would give the same error message, but in the first case the error was never
cleared. This has now been fixed. However, it is possible that there may be
some more errors of this type lurking in the parser. Some time I hope to
systematically look for any other occurrences.

Also I eliminated one new line from error messages taking the form:

-----
fatal:syntax error: line 12 in file proc1.ncl before or near \n
lst_var()
---------^

fatal:syntax error: possibly an undefined procedure
-----

Formerly there were 2 blank lines following the '-----^' line. Now there is
only one. I considered eliminating the other blank line as well, but thought
maybe I should get some feedback on that question first.


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@9117 10fe2890-8a40-4696-8b1d-7ccbb19c398e