Mostly I made cosmetic changes since the underlying Fortran routine is so convoluted, and I wanted to make sure that when I come back to this function months later, I'll know what's going on.
The rip_cape_3d seems to be okay, but I'm not sure about rip_cape_2d.
I haven't been given any tests to run and compare results, so I'm going
based on hoping my interface routine is perfect.  Given the complexity
of this routine, I doubt it is bug free.


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@9118 10fe2890-8a40-4696-8b1d-7ccbb19c398e