This fixes the fatal:syntax error: line -1 before or near whatever() ----------^
message that occurred for script files that do not have a final
newline character.
It is no longer an error and no message appears.
tested on lightning, meeker, tempest and my mac.
Note I can't test on jupiter because I don't have permissions to
modify the source.
FYI on other systems besides my mac, I am testing by going to the
$NCARG/ni/src/ncl directory and moving the symbolic links for the
files I want to modify out of the way and then copying in the files,
compiling, performing the tests, and then moving the links back.
I can't do that on jupiter because there are no directories with symbolic
links.


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@9143 10fe2890-8a40-4696-8b1d-7ccbb19c398e