Added '-e <ext>' option: if specified, input files without extensions that specify their data type (ex. 'foo.grb' or 'bar.hdf') have <ext> applied (as '.ext') so that NCL's addfile() and addfiles() functions will properly open them.
According to Dennis this is especially useful when obtaining a large number
of files (typically GRIB data) that do not have an extension to indicate their
type.  Example usage:

    ncl_convert2nc foo bar baz -e grb

Note that if an input file already has an extension that indicates its type,
the '-e <ext>' will not apply to it, even if specified.  This allows for the
continued mixing of files with differing data formats as arguments to the tool.
For example:

    ncl_convert2nc foo.hdf bar baz -e grb

will treat 'foo' as HDF data, and 'bar' and 'baz' as GRIB data.

Added functionality so that files that exist without an extension to indicate
their type may be specified as having an extension.  Example:

    [file 'ruc' exists on disk, and is GRIB data]
    ncl_convert2nc ruc.grb

The tool will specify 'ruc.grb' as its input data, and NCL's addfile() and
addfiles() functions will act as noted above, and properly process the file.
This was done to maintain functionality that ncl_grib2nc (deprecated) had.
Note that the user must know what type of data they are working with.

Also added writing of global attributes contained in HDF files to the resultant
netCDF output files.


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@9160 10fe2890-8a40-4696-8b1d-7ccbb19c398e