The GRIB 1 reader has been modified to handle varying time units for variables that have records with different time unit indicators (see http://www.nco.ncep.noaa.gov/pmb/docs/on388/table4.html).
If in parsing a record that belongs to a particular variable it encounters
a time unit indicator different from the established indicator, it
compares them and chooses the one with the shortest duration to be the
common unit.
Then it sets the time period for each record based on converting it into
the common units. For periods of a month or greater, this is currently
not entirely accurate, because it is based on a year of 365.25 days and
the average number of hours in a month. However, because no files of this
type have yet been encountered (and may never be), it does not seem
worthwhile to develop anything more accurate at this point. The code does
print warnings (probably too many) if this situation is encountered.


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@9210 10fe2890-8a40-4696-8b1d-7ccbb19c398e