This revises the the GRIB1/GRIB2 differentiation code to fix a problem that resulted in PyNIO not being able to read any GRIB files.
It also moves GribAddFileFormat to AddFileFormats.c and changes the name
of GribVersion to _NclGribVersion in recognition of its external character.
This is to preserve the top level NclFile classes ignorance of the
underlying formats as much as possible.
There is still a global variable called grib_version. It is now declared
in Formats.c, where _NclGribVersion resides.

NclGRIB.c: restores more code that was lost with the GRIB2 enabling check-in.

AddFileFormats.c: grib_version, GribRec, and Grib2Rec are now declared
extern here. (For the time being the NclGRIB2.h is included for the definition
of Grib2Rec, but after NclGRIB.c is included in the build that will not be
necessary). GribAddFileFormat is moved from NclFile.c to here.

NclFile.c: the GRIB header files no longer need to be included.
grib_version is declared extern and there is code that calls _NclGribVersion
if a GRIB extension is encountered.


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@9244 10fe2890-8a40-4696-8b1d-7ccbb19c398e