This "commit" was only intended to update "cpex16.f".  I don't know why the other two changed; I'll have to look into that.  Perhaps I forgot to check in some earlier changes, but I don't remember any such.
At any rate, I am checking in code to fix the same problem that caused the
odd artifacts on plot "B" in Tim's document "POP PyNGL Visualization Timings".
Dave Brown will have to complete the job, however, as he took code from this
example, translated it into C (?), and embedded it elsewhere ...

The basic problem was that the routines ICEGSP and ICEGDP, which are supposed
to tell you where a specified point on the surface of a sphere is relative to
the boundary of a "grid" of points on the surface both contain a somewhat
heuristic test that fails when the "grid" becomes too large and this caused
problems for the algorithm in the version of CPMPXY embedded in "cpex16.f".
I probably ought to add some commenting to ICEGSP and ICEGDP to warn people
who adapt the routine for other purposes that it has some limitations.


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@9289 10fe2890-8a40-4696-8b1d-7ccbb19c398e