Updates for TIGGE-specific parameters.  Some parameters have a description, and more importantly, a short name, that differ from other GRIB v2 data centers.  These parameters are now current with TIGGE's specifications.
NOTE: There are some parameters used by TIGGE that overlap; that is, they are
of the same product discipline and category, and share the same parameter
number.  This is the result of TIGGE differentiating variables based on level
parameters (so far, it appears that nobody else does this).  An example would
be the 'Pressure Level Parameter' U-velocity (defined as "Zonal (Eastward)
wind with a vertical coordinate on a pressure level."), and the 'Single Level
Parameter' 10 meter U-Velocity (defined as "Zonal (Eastward) wind with a
vertical coordinate in a height of 10 m.").  They are both Discipline=0,
Parameter Category=2, and Parameter Number=2.

Given this, it is possible to have distinct parameters that would share the
same entry in a table.  Without a way to differentiate the uniqueness of a
parameter, some parameters are "not found" when consulting code tables at runtime.

It should be noted that there are ways to differentiate such parameters, but
implementing this for a robust long-term solution requires more attention.

In the interim, we will use a "TIGGE-specific" definition module (thank you to
DSS and Doug Schuster for the definition) to handle these cases.


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@9336 10fe2890-8a40-4696-8b1d-7ccbb19c398e