Updates the routine CTTMRG to make it much more efficient for large input arrays (such as the one that Tim Scheitlin was doing PyNGL testing with).
In order to make this routine more efficient, I have had to impose one
additional condition on the user-provided routine RTMI: It must be the
case that IINO is returned less than or equal to the input value of IINI
and that JINO is returned less than or equal to the input value of JINI.
This has been the case for every version of this routine that I have written
(for use in the examples "ctllg[1-3]", "ctpopg", "ctorca", "ctfite", "ctgaus",
"ctswth", "cttd01", and "cttd02") and I think it is always possible to satisfy
the condition.

This fix will not change the timing of CTTMRG in the context of the PyNGL
program that Tim used because Dave Brown's HLU code to call CTTMRG does not
do so in such a way as to create a truly seamless triangular mesh; instead,
it leaves the edges corresponding to first indices I=1 and I=3601 unjoined,
thus avoiding the section of CTTMRG that was so inefficient.  Whether this
should be "fixed" or not is open to debate: the visible effect of leaving a
"seam" on such a dense mesh is generally negligible and closing it involves
some overhead, because one must copy the MxN input arrays to larger arrays
(in this case, from 3600x2400 arrays to 3601x2400 arrays), which is somewhat
expensive in terms of both memory and time used.


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@9348 10fe2890-8a40-4696-8b1d-7ccbb19c398e