Made initial round of changes to WRF routines as recommended by Cindy Bruyere.
The main change is for functions that take Qv (water vapor mixing ratio)
as input; Qv is set to 0 at all locations where Qv is < 0. This
is for wrf_rh, wrf_slp, and wrf_td.

The other big change is that wrf_smooth_2d now only has two arguments
instead of three (the second argument apparently was just a copy of
the first).


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@9378 10fe2890-8a40-4696-8b1d-7ccbb19c398e