The hyi2hyo function has been updated to look for a missing value attached to the input array "xi", and pass this to the Fortan routine. If one is not attached, then 1e20 is used. Upon return, the C interface checks if "intflag" is equal to -1. If so, then this means there are missing values in the return data, and hence a _FillValue attribute needs to be attached to the return variable.
A new function eofunc_varimax_msg has been created so Dennis can test
a missing value version of "eofunc_varimax". (This function should
eventually just replace eofunc_varimax, once we verify that it
works). The interface is identical to the one for eofunc_varimax,
except the input variable is checked to see if it has a _FillValue
attribute, and if so, it is passed to the Fortran routine. If no
_FillValue attribute is attached, then the default for the type of the
input variable is passed in (-999 for integers and floats, -9999 for
doubles, Dennis is this okay?).


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@9385 10fe2890-8a40-4696-8b1d-7ccbb19c398e