This adds a new enumeration type that allows the higher level NclFile class to distinguish between formats. An element of the NclFile object holds the type which gets initialized by the low-level InitializeFormatRec function (declared in NclFileInterfaces.h). This allows the high level function to distinguish between e.g. the GRIB and GRIB2 modules (which allow the same suffixes and therefore cannot be distinguished in the traditional way). This is only needed in very limited circumstances relating to setting format-specific options.
The parameter needed to be added for all the formats.

GRIB2: cleaned a lot of cruft out of the Grib2Rec structure as a starting
point for adding new data structures for handling multiple grid templates.
Modified handling of the "grid_type" attribute for the case of Quasi-Regular
(thinned) grids. The attribute value now dynamically adapts to mention
the type of interpolation (linear or cubic) used to expand the data.
This was what required the changes at the NclFile class level -- because
the attribute values are stored both within the specific format code and
by the higher level class.


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@9406 10fe2890-8a40-4696-8b1d-7ccbb19c398e