Fix for problem that manifests itself on AIX.  When a user has data files with no suffixes, and specifies the "-e" option to denote a file suffix (such as "grb" or "hdf"), the following test
    set sfxlen = `echo $fil:e | awk '{print length($0)}'`

would set the value of 'sfxlen' to a zero-length string (i.e. "").  The shell
pathname modifier ':e' finds no delimiter (default '.') on which to break up
the pathname and thus sends a zero-length string to awk, which does nothing.

The check

    if ($sfxlen == "") then
        set sfxlen = 0
    endif

handles this situation and sets an appropriate value for 'sfxlen'


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@9417 10fe2890-8a40-4696-8b1d-7ccbb19c398e