Makes the test fields used in these examples, which are generated using a random number generator, the same, no matter what machine you run the example on.  This is to address an issue raised by Mary.
I think, however, that the plots you get will still vary from machine
to machine, because a different random number generator is used inside
the VASPACKT routines that position simple vectors, curly vectors, and
streamlines.  (The same field will be depicted now, but the details of
the depiction may vary.)

I am reluctant to change the internal random number generator.  I will
have to think about the ins and outs of this.


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@9484 10fe2890-8a40-4696-8b1d-7ccbb19c398e