Fixes a problem that manifested in the "cropped" DWD GRIB 1 files from Wenchieh Yen: NCL was not distinguishing parameters that had the same parameter number but belong to different parameter tables.
It also relaxes the condition for distinguishing variables via the
'time range indicator', now allowing indicator types 0 and 1 to belong
to the same variable. This change
does not cause any problems with any of the sample GRIB 1 files that I have
on hand. And it fixes the problem reported by Joey Comeaux for ECMWF files
such as Y87469 (I will make available).

However, there is a small possibility that some GRIB files may be read
incorrectly because of this change. A none-definitive symptom will be that
NCL will report 'duplicated' records (See below).

I also added a warning message, similar to what exists for GRIB 2, when
an apparently duplicate GRIB record is encountered. The record numbers of
the 2 'duplicates' are printed and the number of the one that will be ignored.
Duplicate records may truly be duplicates, or they may be a sign that NCL
is not able to distinguish between 2 records that are actually different.
The only way to know is to look at the duplicated records individually with
a record-oriented tool such as wgrib.


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@9560 10fe2890-8a40-4696-8b1d-7ccbb19c398e