Added a gsnp_write_debug_info procedure that is called from all the gsn_xxxx major plotting functions.
It writes the data being plotted and the plot attributes being set to
a netCDF file. It also creates an NCL script that reads in this data,
sets the plot options, and generates the identical plot.

This is for plotting debug purposes only. If somebody is having
problems with a gsn_xxxx plotting script, then they can set the
gsnDebugWriteInfo resource to True, rerun the script, and then just
send us the new debug information that gets created (the netCDF file
and the NCL script). This way they don't have to send us a big NCL
script that may do a bunch of data processing, or the original data
file (if it is large).

By default, the script will be called "debug.ncl", the data "debug.nc",
and the resource file "debug.res" (this resource file is basically
identical to the ".hluresfile" that we ask people to use).
If you want to change "debug" to something else, set
gsnDebugWriteFileName.


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@9672 10fe2890-8a40-4696-8b1d-7ccbb19c398e