This check-in adds a NETCDF_DEBUG define option (set at the top of the file). If set to 1, all NetCDF library calls are printed on stdout.
Also changes all the ncopen calls to nc__open with a big buffer size 'hint'
for the extra parameter.
This was requested quite some time ago by John Clyne to improve
performance. I had actually tried to do this awhile ago, but it turned
out the 'SuppressClose' and the 'DefineMode' options were causing the open
to happen in other places in the code.


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@9684 10fe2890-8a40-4696-8b1d-7ccbb19c398e