AddBuiltIn.c BuiltInFuncs.c Adds a procedure 'status_exit' that takes an integer argument and passes it to the C library 'exit' function, such that a status can be returned to the command line. Note that only the lower 8 bits of the value are returned.
Ncl.c
The cleanup functions NhlClose and _NclFileCleanup are setup using the C library
'atexit' function, which causes them to be called whenever NCL does a normal
exit. This fixes the lack of cleanup in the original NCL 'exit' procedure.

Symbol.c
Modification to the _NclFileCleanup function: the function no longer
returns when a NULL file object is encountered; it simply skips over that
object and continues to look for other possible file objects that might exist.

NclData.c
Changes the management of NCL's fundamental list of current objects to avoid
the problem of the object id value exceeding MAX_INT for long running processes.
Now as objects are deleted, their ids are put on a free list (if there is room).
When new objects are created, an id is taken from the free list if available --
the current maximum id increases only when the free list is empty. Also
increased the object hash table size by a factor of 8 to make object searches
quicker. These changes seemed to result in about a 60% speedup in the complex
script that
motivated this work.


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@9795 10fe2890-8a40-4696-8b1d-7ccbb19c398e