This fixes a problem discovered by Mary Haley and/or Dave Brown.  The routine MDRGSX, which is one of two routines used by MDRGOL and MDRGSF to draw RANGS map outlines and fill RANGS map polygons in a single 1-degree grid box (the other being MDRGSQ), plus the routines MDRGDP, MDRGFA, and MDRGFP, which are called as a result of a call to MDRGSX, were all using SET, GETSET, and MDPRS calls in a manner that prevented a user (in this case, Dave Brown) from overriding a SET call done by MAPINT in order to, for example, move EZMAP graphical output to a different position on the plotter frame and clip it appropriately.
In the process of fixing this bug, I also fixed a couple of incorrect
comments and added a couple of other comments to clarify the purpose of
some lines of code that are otherwise difficult to grok.

I have tested these fixes in the original context by making a version of
NCL containing them and using it to process the original ".ncl" file.  I
have not yet rerun my own tests of the new routines; I do not anticipate
problems, but I will run those tests next, just to be sure.


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@9849 10fe2890-8a40-4696-8b1d-7ccbb19c398e