Updated to use latest LPGL version of the ECMWF code formerly in QU2REG.f.
The new version is based on qu2reg3 rather than qu2reg2. I left the files
as separate files rather than combining 3 files in one as QU2REG.f had done.
The reasoning is to make it look as much as possible like the original files,
not to say this is important...


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@9850 10fe2890-8a40-4696-8b1d-7ccbb19c398e