Christine Shields was running into a problem with paleo_outline where she got the following error message:
 SVBLED - INTEGER WORKSPACE OVERFLOW

It turns out an internal Fortran integer work array was hard-coded to
2000, which isn't big enough. I changed this to a dynamic array that
is allocated in the C wrapper, and then passed to the Fortran routine.

It is not clear to me what the size of this array should be, so I made
it as big as the "zdat" array (or 2000 if zdat is smaller).


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@9863 10fe2890-8a40-4696-8b1d-7ccbb19c398e