This fixes some bugs related to the modification of the object list management in order to handle long-running scripts that create more than 2^31 objects during their lifetime. The original scheme that kept track of recently destroyed objects and recycled their ids was subject to weird failures because NCL in a number of situations uses the _NclGetObj routine to determine if an object still exists. When recently destroyed object ids are reused this can cause NCL to think objects still exist when the id actually belongs to a different newly created object.
The solution was to redesign the scheme (and actually simplify it substantially).
Since the range of acceptible ids is very large (2^31), now NCL simply goes
through the whole ID space, and when it reaches the limit, resets the counter
to a value much larger than 0, near where the long-lived objects have ids, but
much less than the limit. This means that recently released ids are not going to
be used until a long time in future of the program. Currently the counter
resets to 100,000,000. This has been tested with the full test-suite and found to
produce results identical to a version of NCL built with the original object id
management scheme.

Note that at one point in considering how to fix this problem I considered changing
the type of the id struct member to 'long long' (i.e. a 64-bit long) so that the
adaptation above would not be necessary, but it turned out it would have resulted
in a long cascasde of other changes to fundmental data structures that would have
taken a very long time to reliably test and verify. The route I eventually
followed now seems to be such an obvious choice that I should have immediately
realized it. Oh well...


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@9914 10fe2890-8a40-4696-8b1d-7ccbb19c398e