This check-in adds support for GRIB 2 files from NDFD. These files are available from http://www.weather.gov/ndfd/anonymous_ftp.htm.
These files have 2 unusual features: 1) they supply an explicit missing
value (it always seems to be 9999.0) and 2) the grids rows alternate
directions: in order to read them you need to reverse the values of
every other row -- g2clib does not do this for you.

Jennifer Adams, the GrADS developer, was kind enough to bring these files to
my attention.

There is one type of file from this agency that is still not read completely:
the 'WX' files. These are the first GRIB 2 files that we have encountered
that have a 'local' usage section. The values in the data grid are integers
that represent offsets into an array of character stuff in the local usage
section -- which in turn are codes that mean things like 'partly cloudy' or
'scattered snow showers'. This would be quite involved to decode and there's
no time for it now.


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@9945 10fe2890-8a40-4696-8b1d-7ccbb19c398e